[W.I.P] Keep the existing UNALIGNED macro structure
The below statement make a build error with  "-Werror -Wignored-attributes"
because Clang does not have an attribute that is used like the below that.
 *(int __attribute__((aligned(1))) *)ptr = 2; // Wrong, attribute ignored

In this case, we need to use " -Wno-error=ignored-attributes" or
"-Wno-ignored-attributes" flag to avoid the build error.

Or, we need to use a typedef such as the below statement.
typedef int __attribute__((aligned(1))) unaligned_int;
    *(unaligned_int *)ptr = 2; // Works

However, as far as we are aware there is no Clang drop-in replacement
for the UNALIGNED macro used by CoreCLR. It looks like the trial to
modify the existing UNALIGNED statements would be a tedious change to us too.

Signed-off-by: Geunsik Lim <geunsik.lim@samsung.com>